Add option flag to cd (Call DLL) for better perf
From the "J for C Programmers" -- 21. Calling External Programs

  > ### Options
  >
  > *Flag characters* that control the call to DLL may be included just before the first descriptor.  The `+` and `%` flags control technical details of the calling convention and will not be discussed here.  The `>` flag indicates that you are not interested in anything but the result from the DLL.  When `>` is given, the result of cd is an unboxed scalar rather than a list of boxes, which can be much faster.
  >
  > The cd verb has rank `1 1`.  If you have to call the same entry point many times, it may be faster to have an array of arguments and use `cd` once (preferably with the `>` flag).

I found that foreign functions that return a non-atomic type (eg. `*c`)
do not like the `>` option, so for those I'm still using `0 {::`

Also, as `cd` is rank `1 1`, I removed an explicit `"1` on a repeated
call to `DrawCircle` in the `examples/shoot.ijs`.